Moved `logic_thread` from autonomous to logic module
